backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
opencv-python
Pillow
pydantic
python-dotenv
google-generativeai
transformers
scikit-learn
pandas
matplotlib
